Introduce tests: Add test for AWS JSON protocol conflicts
This change is prompted by issues in other projects that _use_ the
content-api-firehose-client - in the `apple-news` project, updating another
library transitively updated the `aws-json-protocol` to a version that
was incompatible with the AWS SDK artifacts bought in by
content-api-firehose-client - leading to runtime errors:

https://github.com/guardian/apple-news/pull/386#issuecomment-2501369378

Here we introduce an integration test that exercises the AWS SDK
`DescribeStreamSummary` call that failed - it checks that the response
can be successfully processed by the AWS SDK. Note that although we have
introduced this test, and it _passes_, this doesn't mean we've fixed the
issue with other projects yet - it just allows us to reproduce the issue,
by bumping the version on `aws-json-protocol` (ie by adding
`"software.amazon.awssdk" % "aws-json-protocol" % "2.29.23"` to the deps),
which we do outside of this branch.

In order to get the test running in CI, we have to allow the GitHub Actions
workflow to make the AWS SDK `DescribeStreamSummary` call, which requires
AWS credentials.

We need a IAM role that we can use in the GitHub Actions runners to
have access to the kinesis stream we’re describing - this is the same
approach taken in `facia-scala-client` with
https://github.com/guardian/facia-scala-client/pull/272 .

This commit copies the example cdk stack from facia-scala-client,
updating the version of (gu)cdk used to the latest version, and tweaks
the role policies to match what we need to do.

This includes a fix from Akash to get Region and Account - they were `undefined`
before! As Akash says:

> scope here is of type App. The region and account properties of scope will not be meaningful here as we're not using CDK's Stage construct. That is, they'll return undefined.

https://github.com/guardian/content-api-firehose-client/pull/51#discussion_r1870959622