Tell sbt aws-json-protocol uses pvp versioning
Recently, we ran into a problem because one dependency was depending on
version 2.29.x of aws-json-protocol while another depended on 2.25.x:
when our app (apple-news) tried to communicate with kinesis, it failed
to parse the json responses properly and threw errors.

In an attempt to get sbt to warn us about this situation, this commit
sets
[libraryDependencySchemes](https://www.scala-lang.org/blog/2021/02/16/preventing-version-conflicts-with-versionscheme.html#versionscheme-librarydependencyschemes-and-sbt-150)
to “[pvp](https://pvp.haskell.org/)” for aws-json-protocol: once this is
done, sbt errors when a 2.25.x dependency is evicted by a 2.29.x
dependency (as pvp considers the first two digits in a version number to
specify the major version).

This is not quite a match for [how AWS describe the version policy for
these
libraries](https://github.com/aws/aws-sdk-java-v2/blob/master/VERSIONING.md),
but it’s close enough for this purpose: the main problem is that minor
versions can include backwards-incompatible changes, so we probably want
sbt to warn us or error when evicting for a bigger minor version.

Do we want to do this? We’d need to do it across all our repositories,
for any aws sdk dependencies, which seems like a lot of overhead. Also,
it might well lead to false positives: two versions A.B and A.C are not
_guaranteed_ to be incompatible, and we’re probably in this situation in
places already.